+++ MeepMeepTesting/src/main/java/com/example/meepmeeptesting/meepMeepCode.java
@@ -3,8 +3,6 @@ package com.example.meepmeeptesting;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
-import com.acmerobotics.roadrunner.trajectory.constraints.MinVelocityConstraint;
-import com.acmerobotics.roadrunner.trajectory.constraints.TrajectoryVelocityConstraint;
 import com.noahbres.meepmeep.MeepMeep;
 import com.noahbres.meepmeep.core.colorscheme.scheme.ColorSchemeBlueLight;
 import com.noahbres.meepmeep.roadrunner.DefaultBotBuilder;
@@ -18,96 +16,29 @@ public class meepMeepCode {
 
         MeepMeep meepMeep = new MeepMeep(800);
         RobotConfig robot = new RobotConfig();
-        robotConstants RobotConstants = new robotConstants();
+//        RobotConstants RobotConstants = new RobotConstants();
 
         RoadRunnerBotEntity myBot = new DefaultBotBuilder(meepMeep)
                 .setConstraints(40, 45, Math.toRadians(200), Math.toRadians(200), 10.64)
                 .setColorScheme(new ColorSchemeBlueLight())
 
                     .followTrajectorySequence(drive -> (drive.trajectorySequenceBuilder(new Pose2d(-35, 62, Math.toRadians(270)))).setVelConstraint(SampleMecanumDrive.getVelocityConstraint(45, Math.toRadians(310), 10.62))
-                            .splineToConstantHeading(new Vector2d(-51, 37), Math.toRadians(270))
-                            .addTemporalMarker(.1, () -> {
-                                robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
-                            })
-                            .splineToLinearHeading(new Pose2d(-47, 34.3), Math.toRadians(30))
-                            .waitSeconds(.6)//.4
-                            .addTemporalMarker(3.0, () -> {
-                                robot.rightPixelServo.setPosition(RobotConstants.rightIn);
-                            })
-                            .addTemporalMarker(2.9, () -> {
-                                robot.intakeMotor.setPower(1);
-                                robot.transferMotor.setPower(1);
-                                robot.stackArm(RobotConfig.StackArm.OUT);
-                            })
-                            .back(.1)
-                            .splineToConstantHeading(new Vector2d(-57, 34.9), Math.toRadians(180))
-                            .addTemporalMarker(3.7, () -> {
-                                robot.grabFromStack(1);
-                            })
-                            .waitSeconds(.9)
-                            .forward(.5)
-                            .splineToConstantHeading(new Vector2d(-33, 57), Math.toRadians(0))
-                            .lineTo(new Vector2d(31, 57))
-                            .addTemporalMarker(7.1, () -> {
-                                robot.setTargetSlidePos(RobotConfig.SlideHeight.PRELOAD_DROP);
-                                robot.intakeMotor.setPower(0);
-                                robot.transferMotor.setPower(0);
-                            })
-                            .splineToConstantHeading(new Vector2d(50, 28.1), Math.toRadians(0))
-                            .lineTo(new Vector2d(53, 28.1), RobotConfig.getVelocityConstraint(25, Math.toRadians(200), 10.62), RobotConfig.getAccelerationConstraint(25))
-                            .addTemporalMarker(10.0, () -> {
-                                robot.dropper(RobotConfig.Dropper.OPEN);
-                                robot.safeRelocalizeBackdrop();
-                            })
-                            .addTemporalMarker(10.4, () -> {
-                                robot.setTargetSlidePos(RobotConfig.SlideHeight.LOW);
-                                robot.dropper(RobotConfig.Dropper.PARTIAL);
-                            })
-                            .waitSeconds(1.2)
-                            //Cycle 1
-                            .back(.1)
-                            .splineToConstantHeading(new Vector2d(33, 57), Math.toRadians(180))
-                            .addTemporalMarker(12.2, () -> {
-                                robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
-                            })
-                            .lineTo(new Vector2d(-33, 57))
-                            .splineToConstantHeading(new Vector2d(-52, 36), Math.toRadians(250))
-                            .splineToConstantHeading(new Vector2d(-56.5, 33.4), Math.toRadians(180))
-                            .addTemporalMarker(14.4, () -> {
-                                robot.intakeMotor.setPower(1);
-                                robot.transferMotor.setPower(1);
-                                robot.dropper(RobotConfig.Dropper.CLOSED);
-                                robot.stackArm(RobotConfig.StackArm.FAR_RIGHT);
-                            })
-                            .waitSeconds(1.7)
-                            .addTemporalMarker(15.1, () -> {
-                                robot.stackArm(RobotConfig.StackArm.OUT);
-                            })
-                            .addTemporalMarker(15.7, () -> {
-                                robot.grabFromStack(2);
-                            })
-                            //To backdrop
-                            .forward(.5)
-                            .splineToConstantHeading(new Vector2d(-33, 57), Math.toRadians(0))
-                            .lineTo(new Vector2d(31, 57))
-                            .addTemporalMarker(20.8, () -> {
-                                robot.setTargetSlidePos(15);
-                                robot.intakeMotor.setPower(0);
-                                robot.transferMotor.setPower(0);
-                            })
-                            .splineToConstantHeading(new Vector2d(50, 37), Math.toRadians(0))
-                            .lineTo(new Vector2d(53, 37), RobotConfig.getVelocityConstraint(25, Math.toRadians(200), 10.62), RobotConfig.getAccelerationConstraint(25))
-                            .addTemporalMarker(22.2, () -> {
-                                robot.dropper(RobotConfig.Dropper.PARTIAL);
-                                robot.safeRelocalizeBackdrop();
-                            })
-                            .waitSeconds(1)
-                            .addTemporalMarker(23.8, () -> {
+                            .lineTo(new Vector2d(12, 45))
+                            .splineToConstantHeading(new Vector2d(8.5, 36), Math.toRadians(180))
+                            .lineTo(new Vector2d(8, 36))
+                            .addTemporalMarker(1.5, () -> robot.rightPixelServo.setPosition(RobotConstants.rightIn))
+                            .lineTo(new Vector2d(20, 36))
+                            .splineToSplineHeading(new Pose2d(45, 31, Math.toRadians(0)), Math.toRadians(0))
+                            .lineTo(new Vector2d(53, 31))
+                            .addTemporalMarker(4.1, () -> robot.dropper(RobotConfig.Dropper.OPEN))
+                            .waitSeconds(.3)
+                            .lineTo(new Vector2d(40, 31))
+                            .addDisplacementMarker(() -> {
                                 robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
                                 robot.dropper(RobotConfig.Dropper.CLOSED);
                             })
-                            .back(6)
-                            .waitSeconds(1.5)
+                            .lineTo(new Vector2d(40, 12))
+                            .lineTo(new Vector2d(45, 12))
                             .build());
