--- MeepMeepTesting/src/main/java/com/example/meepmeeptesting/robotConstants.java
@@ -1,7 +1,0 @@
-package com.example.meepmeeptesting;
-
-public class robotConstants {
-
-    public static  double leftIn = 0.5, rightIn = .5;
-
-}
