+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/fCamRedCloseAuto.java
@@ -14,6 +14,7 @@ import org.firstinspires.ftc.teamcode.drive.RobotConfig;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.pipelines.ColorMask;
 import org.firstinspires.ftc.teamcode.robot.HwMap;
+import org.firstinspires.ftc.teamcode.robot.PassData;
 import org.firstinspires.ftc.teamcode.robot.RobotConstants;
 import org.firstinspires.ftc.teamcode.robot.RobotMethods;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
@@ -52,8 +53,12 @@ public class fCamRedCloseAuto extends LinearOpMode {
         RobotConfig robot = new RobotConfig(hardwareMap);
         ColorMask pipeline = new ColorMask();
 
-//        robot.leftLiftServo.setPosition(intakePos+RobotConstants.stackLeftOffset);
-//        robot.rightLiftServo.setPosition(intakePos);
+        telemetry.setAutoClear(false);
+        Telemetry.Item detectedPos = telemetry.addData("Position", "No detection");
+        Telemetry.Item IMU = telemetry.addData("Current IMU", robot.getCurrentIMU().toString());
+        Telemetry.Item Park = telemetry.addData("Park Position", PassData.parkPosition.toString());
+        telemetry.update();
+
         robot.transferMotor.setPower(-.2);
         robot.stackHold(false);
 
@@ -92,8 +97,8 @@ public class fCamRedCloseAuto extends LinearOpMode {
                     robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
                     robot.dropper(RobotConfig.Dropper.CLOSED);
                 })
-                .lineTo(new Vector2d(40, -12))
-                .lineTo(new Vector2d(45, -12))
+                .lineTo(new Vector2d(40, PassData.parkPosition.red))
+                .lineTo(new Vector2d(45, PassData.parkPosition.red))
                 .build();
 
         TrajectorySequence center = robot.trajectorySequenceBuilder(new Pose2d(12, -63, Math.toRadians(90)))
@@ -111,8 +116,8 @@ public class fCamRedCloseAuto extends LinearOpMode {
                     robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
                     robot.dropper(RobotConfig.Dropper.CLOSED);
                 })
-                .lineTo(new Vector2d(40, -12))
-                .lineTo(new Vector2d(45, -12))
+                .lineTo(new Vector2d(40, PassData.parkPosition.red))
+                .lineTo(new Vector2d(45, PassData.parkPosition.red))
                 .build();
 
         TrajectorySequence right = robot.trajectorySequenceBuilder(new Pose2d(12, -63, Math.toRadians(90)))
@@ -130,19 +135,13 @@ public class fCamRedCloseAuto extends LinearOpMode {
                     robot.setTargetSlidePos(RobotConfig.SlideHeight.BOTTOM);
                     robot.dropper(RobotConfig.Dropper.CLOSED);
                 })
-                .lineTo(new Vector2d(40, -12))
-                .lineTo(new Vector2d(45, -12))
+                .lineTo(new Vector2d(40, PassData.parkPosition.red))
+                .lineTo(new Vector2d(45, PassData.parkPosition.red))
                 .build();
 
 
 
         ElapsedTime cameraDelayTimer = new ElapsedTime();
-
-        telemetry.setAutoClear(false);
-        Telemetry.Item detectedPos = telemetry.addData("Position", "No detection");
-        Telemetry.Item wasPos = telemetry.addData("Was pos", "");
-
-//        Telemetry.Item slideData = telemetry.addData("Slide Data:", "Encoder Val:" + robot.liftEncoder.getCurrentPosition() + " Target Val:" + targetSlidePos);
 
         robot.leftPixelServo.setPosition(RobotConstants.leftOut);
         robot.dropServo.setPosition(RobotConstants.dropClosed);
@@ -161,7 +160,9 @@ public class fCamRedCloseAuto extends LinearOpMode {
         robot.setTargetSlidePos(RobotConfig.SlideHeight.PRELOAD_DROP);
 
         while (opModeIsActive() && !isStopRequested()) {
-//            robot.update();
+            robot.update();
+            IMU.setValue(robot.getCurrentIMU().toString());
+            telemetry.update();
 
             switch (camera) {
                 case WAIT:
@@ -189,17 +190,12 @@ public class fCamRedCloseAuto extends LinearOpMode {
                             detectedPos.setValue("Default center (No detection)");
                             break;
                     }
-//                    wasPos.setValue(RobotMethods.updateRobotPosition(robot.getPoseEstimate()));
 
                     camera = Camera.FINISHED;
                     break;
                 case FINISHED:
                     break;
             }
-
-           robot.update();
-
-//            slideData.setValue( "Encoder Val: " + slideCurPos + " Target Val: " + targetSlidePos + " Slide Power: " + (double)Math.round(slidePower*100)/100);
 
         }
