+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/old/TestingBack.java
@@ -1,8 +1,9 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.autos.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.drive.RobotConfig;
@@ -10,8 +11,9 @@ import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.robot.HwMap;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
+@Disabled
 @Autonomous
-public class TestingForwards extends LinearOpMode {
+public class TestingBack extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -25,7 +27,7 @@ public class TestingForwards extends LinearOpMode {
                 .back(70)
                 .build();
 
-        Trajectory forwards = robot.trajectoryBuilder(new Pose2d(0, 0, Math.toRadians(0)))
+        TrajectorySequence forwards = robot.trajectorySequenceBuilder(new Pose2d(0, 0, Math.toRadians(0)))
                 .forward(70)
                 .build();
 
@@ -33,6 +35,6 @@ public class TestingForwards extends LinearOpMode {
         robot.update();
         robot.setPoseEstimate(new Pose2d(0, 0, Math.toRadians(0)));
 
-        robot.followTrajectory(forwards);
+        robot.followTrajectory(backwards);
     }
 }
