+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/old/TestingForwards.java
@@ -1,19 +1,23 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.autos.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.drive.RobotConfig;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.robot.HwMap;
+import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
+@Disabled
 @Autonomous
-public class TestingForwardsOld extends LinearOpMode {
+public class TestingForwards extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        SampleMecanumDrive robot = new SampleMecanumDrive(hardwareMap);
+        RobotConfig robot = new RobotConfig(hardwareMap);
 //        SampleMecanumDrive driveTrain = new SampleMecanumDrive(hardwareMap);
 //        HwMap robot = new HwMap();
 //        robot.init(hardwareMap);
