+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/old/TestingForwardsOld.java
@@ -1,16 +1,17 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.autos.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.drive.RobotConfig;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
-import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
+@Disabled
 @Autonomous
-public class TestingBackOld extends LinearOpMode {
+public class TestingForwardsOld extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -24,7 +25,7 @@ public class TestingBackOld extends LinearOpMode {
                 .back(70)
                 .build();
 
-        TrajectorySequence forwards = robot.trajectorySequenceBuilder(new Pose2d(0, 0, Math.toRadians(0)))
+        Trajectory forwards = robot.trajectoryBuilder(new Pose2d(0, 0, Math.toRadians(0)))
                 .forward(70)
                 .build();
 
@@ -32,6 +33,6 @@ public class TestingBackOld extends LinearOpMode {
         robot.update();
         robot.setPoseEstimate(new Pose2d(0, 0, Math.toRadians(0)));
 
-        robot.followTrajectory(backwards);
+        robot.followTrajectory(forwards);
     }
 }
