+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/old/fCamRedFarAutoPlaceOnly.java
@@ -1,8 +1,9 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.autos.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -19,8 +20,9 @@ import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
 
-@Autonomous(name = "fCam Blue Far Auto Place Only")
-public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
+@Disabled
+@Autonomous(name = "fCam Red Far Auto Place Only")
+public class fCamRedFarAutoPlaceOnly extends LinearOpMode {
 
     enum Camera {
         WAIT,
@@ -62,7 +64,7 @@ public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
         {
             public void onOpened()
             {
-                pipeline.setAlliance("Blue");
+                pipeline.setAlliance("Red");
                 robot.webcam.setPipeline(pipeline);
 
                 robot.webcam.startStreaming(640,480, OpenCvCameraRotation.UPSIDE_DOWN);
@@ -74,73 +76,73 @@ public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
             }
         });
 
-        TrajectorySequence left = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+        TrajectorySequence right = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, -63, Math.toRadians(90)))
                 .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
-                .lineTo(new Vector2d(-35, 60))
-                .splineToConstantHeading(new Vector2d(-31, 34), Math.toRadians(280))
-                .addSpatialMarker(new Vector2d(-31, 34), () -> robot.leftServo.setPosition(RobotConstants.leftIn))
+                .lineTo(new Vector2d(-35, -40))
+                .splineToConstantHeading(new Vector2d(-31, -34), Math.toRadians(100))
+                .addSpatialMarker(new Vector2d(-31, -34), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(-36, 32))
-                .splineToSplineHeading(new Pose2d(-40, 20, Math.toRadians(0)), Math.toRadians(270))
-                .splineToConstantHeading(new Vector2d(-32, 14), Math.toRadians(0))
-                .lineTo(new Vector2d(-12, 14))
-//                .waitSeconds(1)
-//                .lineTo(new Vector2d(10, 14))
-//                .splineToConstantHeading(new Vector2d(45, 42), Math.toRadians(0))
-//                .addSpatialMarker(new Vector2d(25, 22), () -> targetSlidePos = RobotConstants.slideAuto)
-//                .lineTo(new Vector2d(54.5, 42))
-//                .lineTo(new Vector2d(54.6, 42))
-//                .addTemporalMarker(6.8 +1, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .lineTo(new Vector2d(-38, -33))
+                .splineToSplineHeading(new Pose2d(-41, -20, Math.toRadians(0)), Math.toRadians(90))
+                .splineToConstantHeading(new Vector2d(-32, -10.5), Math.toRadians(0))
+                .lineTo(new Vector2d(-12, -10.5))
+//                .waitSeconds(15)
+//                .lineTo(new Vector2d(10, -10.5))
+//                .splineToConstantHeading(new Vector2d(45, -54), Math.toRadians(0))
+//                .addSpatialMarker(new Vector2d(25, -22), () -> targetSlidePos = RobotConstants.slideAuto)
+//                .lineTo(new Vector2d(54.5, -54))
+//                .lineTo(new Vector2d(54.6, -54))
+//                .addTemporalMarker(7.2+15, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
 //                .waitSeconds(.6)
-//                .lineTo(new Vector2d(40, 42))
+//                .lineTo(new Vector2d(40, -54))
 //                .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-//                .lineTo(new Vector2d(40, 10))
-//                .lineTo(new Vector2d(45, 10))
+//                .lineTo(new Vector2d(40, -16))
+//                .lineTo(new Vector2d(50, -16))
                 .build();
 
-        TrajectorySequence center = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+        TrajectorySequence center = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, -63, Math.toRadians(90)))
                 .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
-                .lineTo(new Vector2d(-35, 60))
-                .splineToSplineHeading(new Pose2d(-34, 17, Math.toRadians(0)), Math.toRadians(280))
-                .addSpatialMarker(new Vector2d(-34, 17), () -> robot.leftServo.setPosition(RobotConstants.leftIn))
+                .lineTo(new Vector2d(-35, -60))
+                .splineToSplineHeading(new Pose2d(-36, -17, Math.toRadians(0)), Math.toRadians(100))
+                .addSpatialMarker(new Vector2d(-36, -17), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(-34, 16.5))
-                .splineToConstantHeading(new Vector2d(-12, 13), Math.toRadians(0))
-//                .waitSeconds(1)
-//                .lineTo(new Vector2d(10, 13))
-//                .splineToConstantHeading(new Vector2d(45, 32), Math.toRadians(0))
-//                .addSpatialMarker(new Vector2d(26, 22), () -> targetSlidePos = RobotConstants.slideAuto)
-//                .lineTo(new Vector2d(55.5, 32))
-//                .lineTo(new Vector2d(55.6, 32))
-//                .addTemporalMarker(6.3+1, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .lineTo(new Vector2d(-36, -16.5))
+                .splineToConstantHeading(new Vector2d(-12, -12), Math.toRadians(0))
+//                .waitSeconds(15)
+//                .lineTo(new Vector2d(10, -12))
+//                .splineToConstantHeading(new Vector2d(45, -42), Math.toRadians(0))
+//                .addSpatialMarker(new Vector2d(26, -22), () -> targetSlidePos = RobotConstants.slideAuto)
+//                .lineTo(new Vector2d(55.5, -42))
+//                .lineTo(new Vector2d(55.6, -42))
+//                .addTemporalMarker(6.5+15, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
 //                .waitSeconds(.6)
-//                .lineTo(new Vector2d(40, 32))
+//                .lineTo(new Vector2d(40, -42))
 //                .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-//                .lineTo(new Vector2d(40, 10))
-//                .lineTo(new Vector2d(45, 10))
+//                .lineTo(new Vector2d(40, -16))
+//                .lineTo(new Vector2d(50, -16))
                 .build();
 
-        TrajectorySequence right = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+        TrajectorySequence left = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, -63, Math.toRadians(90)))
                 .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
-                .lineTo(new Vector2d(-35, 60))
-                .splineToSplineHeading(new Pose2d(-45, 21, Math.toRadians(0)), Math.toRadians(280))
-                .addSpatialMarker(new Vector2d(-45, 21), () -> robot.leftServo.setPosition(RobotConstants.leftIn))
+                .lineTo(new Vector2d(-35, -60))
+                .splineToSplineHeading(new Pose2d(-45, -19, Math.toRadians(0)), Math.toRadians(100))
+                .addSpatialMarker(new Vector2d(-45, -19), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(-42, 19))
-                .splineToConstantHeading(new Vector2d(-30, 14), Math.toRadians(0))
-                .lineTo(new Vector2d(-12, 14))
-                .waitSeconds(1)
-//                .lineTo(new Vector2d(10, 14))
-//                .splineToConstantHeading(new Vector2d(45, 23), Math.toRadians(0))
-//                .addSpatialMarker(new Vector2d(38, 36), () -> targetSlidePos = RobotConstants.slideAuto)
-//                .lineTo(new Vector2d(54.5, 23))
-//                .lineTo(new Vector2d(54.6, 23))
-//                .addTemporalMarker(6.3+1, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .lineTo(new Vector2d(-44, -17))
+                .splineToConstantHeading(new Vector2d(-30, -11), Math.toRadians(0))
+                .lineTo(new Vector2d(-12, -11))
+//                .waitSeconds(15)
+//                .lineTo(new Vector2d(10, -11))
+//                .splineToConstantHeading(new Vector2d(45, -34.5), Math.toRadians(0))
+//                .addSpatialMarker(new Vector2d(38, -36), () -> targetSlidePos = RobotConstants.slideAuto)
+//                .lineTo(new Vector2d(54.5, -34.5))
+//                .lineTo(new Vector2d(54.6, -34.5))
+//                .addTemporalMarker(6.5+15, () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
 //                .waitSeconds(.6)
-//                .lineTo(new Vector2d(40, 23))
+//                .lineTo(new Vector2d(40, -34.5))
 //                .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-//                .lineTo(new Vector2d(40, 10))
-//                .lineTo(new Vector2d(45, 10))
+//                .lineTo(new Vector2d(40, -10))
+//                .lineTo(new Vector2d(45, -10))
 
                 .build();
 
@@ -154,7 +156,7 @@ public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
 
 
 
-        robot.leftServo.setPosition(RobotConstants.leftOut);
+        robot.rightServo.setPosition(RobotConstants.rightOut);
         robot.dropServo.setPosition(RobotConstants.dropClosed);
 
         waitForStart();
@@ -162,7 +164,7 @@ public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
 
         robot.climbMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        driveTrain.setPoseEstimate(new Pose2d(-35, 63, Math.toRadians(270)));
+        driveTrain.setPoseEstimate(new Pose2d(-35, -63, Math.toRadians(90)));
 
         cameraDelayTimer.reset();
 
@@ -197,7 +199,7 @@ public class fCamBlueFarAutoPlaceOnly extends LinearOpMode {
                     }
                     wasPos.setValue(RobotMethods.updateRobotPosition(driveTrain.getPoseEstimate()));
 
-                    driveTrain.setPoseEstimate(new Pose2d(-35, 63, Math.toRadians(270)));
+                    driveTrain.setPoseEstimate(new Pose2d(-35, -63, Math.toRadians(90)));
                     camera = Camera.FINISHED;
                     break;
                 case FINISHED:
