+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/velocityTest.java
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.teamcode.robot;
 
 import android.annotation.SuppressLint;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
@@ -10,6 +11,7 @@ import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.drive.StandardTrackingWheelLocalizer;
 import org.firstinspires.ftc.teamcode.util.Encoder;
 
+@Disabled
 @TeleOp
 public class velocityTest extends LinearOpMode {
     Encoder left, right, center;
